countercase/requirements.txt
pdfplumber>=0.10.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
transformers>=4.30.0
langchain>=0.1.0
langchain-text-splitters>=0.0.1
tiktoken>=0.5.0
pandas>=2.0.0
pyarrow>=14.0.0
tqdm>=4.65.0
faiss-cpu>=1.7.4
torch>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
